+++ demis/applications/secrets.tf
@@ -183,3 +183,23 @@ resource "kubernetes_secret_v1" "rabbit_mq_credentials" {
     RABBITMQ_PASSWORD = var.rabbitmq_password
   }
 }
+
+resource "kubernetes_secret_v1" "ars_secure_queue_encryption_secret" {
+  metadata {
+    name      = "ars-secure-queue-encryption-secret"
+    namespace = var.target_namespace
+    annotations = {
+      checksum = substr(sha256(jsonencode({
+        ARS_SECURE_QUEUE_ENCRYPTION_CURRENT_KEY  = var.ars_secure_queue_encryption_current_secret
+        ARS_SECURE_QUEUE_ENCRYPTION_PREVIOUS_KEY = var.ars_secure_queue_encryption_previous_secret
+      })), 0, 61)
+    }
+  }
+
+  immutable = true
+
+  data = {
+    ARS_SECURE_QUEUE_ENCRYPTION_CURRENT_KEY  = var.ars_secure_queue_encryption_current_secret
+    ARS_SECURE_QUEUE_ENCRYPTION_PREVIOUS_KEY = var.ars_secure_queue_encryption_previous_secret
+  }
+}