+++ dmz/applications/variables_credentials.tf
@@ -40,6 +40,27 @@ variable "ars_bulk_upload_hmac_secret" {
   default     = ""
 }
 
+variable "ars_bis_in_queue_encryption_current_secret" {
+  type        = string
+  sensitive   = true
+  description = "The current encryption key for the bulk upload service. Must be 16 bytes"
+  default     = ""
+}
+
+variable "ars_bis_in_queue_encryption_previous_secret" {
+  type        = string
+  sensitive   = true
+  description = "The previous encryption key for the bulk upload service"
+  default     = ""
+}
+
+variable "ars_secure_queue_encryption_current_secret" {
+  type        = string
+  sensitive   = true
+  description = "The current encryption key for the secure queue. Must be 16 bytes"
+  default     = ""
+}
+
 variable "rabbitmq_username" {
   description = "The RabbitMQ username for the application"
   type        = string